Title: 출석 부르기, Time: 147ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240123/출석 부르기/calling-attendance.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        double a = sc.nextDouble();
+        int a = sc.nextInt();
 
         if (a == 1) {
             System.out.println("John");
